requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.7.14
cffi==1.17.1
channels==4.3.0
charset-normalizer==3.4.2
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
faiss-cpu==1.11.0.post1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.8
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.2
openai==1.97.1
orjson==3.11.1
packaging==25.0
pillow==11.3.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pypdf==5.9.0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.7.31
requests==2.32.4
requests-toolbelt==1.0.0
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
zope.interface==7.2
zstandard==0.23.0
